Player Class and new Player Sprite
Created a class to handle player functions and moved all player related objects and functions to this class. Tidied up the code as well to help make it easier to update and find where things are. Added a new sprite sheet for the player and updated and moved the animation code to reflect this.