Bomb updates and menu system
Added menu system functionality and the ability to transfer from the main menu to the game. Once the player dies, they now move to the game over screen instead of resetting. The player can also return to the main menu or restart the game from the gameover screen. Updated the bomb tile. Created explosion class for explosions and added an animation for them. Once animation is finished, the tile is removed instead of when the timer runs out.