12/12/19 - Menu System, Score System and Obstacle Updates
Added art for all the menus in the game. Implemented these menus into the game. Added the score system with the functionality to update and save the new high score into local storage. Added gem durability, allowing the rarer games to take more bombs to destroy. Added sprites to the gems to show when they are close to breaking.